Changed blocked hashes to a single feature [#1793]
