--- comixed-webui/src/app/comic-pages/actions/upload-blocked-pages.actions.ts
@@ -1,34 +1,0 @@
-/*
- * ComiXed - A digital comic book library management application.
- * Copyright (C) 2021, The ComiXed Project
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses>
- */
-
-import { createAction, props } from '@ngrx/store';
-import { BlockedHash } from '@app/comic-pages/models/blocked-hash';
-
-export const uploadBlockedPages = createAction(
-  '[Upload Blocked Pages] Upload blocked pages',
-  props<{ file: File }>()
-);
-
-export const blockedPagesUploaded = createAction(
-  '[Upload Blocked Pages] Blocked pages uploaded',
-  props<{ entries: BlockedHash[] }>()
-);
-
-export const uploadBlockedPagesFailed = createAction(
-  '[Upload Blocked Pages] Failed to upload blocked pages'
-);
