--- comixed-webui/src/app/comic-pages/selectors/block-page.selectors.ts
@@ -1,27 +1,0 @@
-/*
- * ComiXed - A digital comic book library management application.
- * Copyright (C) 2021, The ComiXed Project
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses>
- */
-
-import { createFeatureSelector } from '@ngrx/store';
-import {
-  BLOCK_PAGE_FEATURE_KEY,
-  BlockPageState
-} from '@app/comic-pages/reducers/block-page.reducer';
-
-export const selectBlockPageState = createFeatureSelector<BlockPageState>(
-  BLOCK_PAGE_FEATURE_KEY
-);
