--- comixed-webui/src/app/comic-pages/selectors/blocked-page-detail.selectors.ts
@@ -1,34 +1,0 @@
-/*
- * ComiXed - A digital comic book library management application.
- * Copyright (C) 2021, The ComiXed Project
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses>
- */
-
-import { createFeatureSelector, createSelector } from '@ngrx/store';
-import {
-  BLOCKED_PAGE_DETAIL_FEATURE_KEY,
-  BlockedPageDetailState
-} from '../reducers/blocked-page-detail.reducer';
-
-/** Loads the blocked page detail feature state. */
-export const selectBlockedPageDetailState =
-  createFeatureSelector<BlockedPageDetailState>(
-    BLOCKED_PAGE_DETAIL_FEATURE_KEY
-  );
-
-export const selectBlockedPageDetail = createSelector(
-  selectBlockedPageDetailState,
-  state => state.entry
-);
