+++ comixed-webui/src/assets/i18n/en/app.json
@@ -123,7 +123,7 @@
     "option": {
       "all-comics": "All comics",
       "batch-processes": "Batch Processes",
-      "blocked-hashes": "Blocked Pages",
+      "blocked-hashes": "Blocked Hashes",
       "build-details": "Build Details",
       "changed-comics": "Changed Comics",
       "character-collection": "Characters",
